Add evaluation form delete, decimal answers, and fix hidden attachments
- Delete an evaluation form (admin only), removing its R2 objects too
- Lock a criteria's inputType once answers exist, so stored values can't be
  orphaned in the column the view no longer reads
- Accept decimal answers: numberValue Int -> Float, drop Zod .int(),
  set step="any" on the number input
- Show attachments on criteria that have no text/number answer, and on
  NUMBER criteria at all — both were previously invisible in the view

+++ server/src/modules/evaluations/evaluations.dto.ts
@@ -6,7 +6,7 @@ export const createEvaluationSchema = z.object({
     .array(
       z.object({
         criteriaId: z.string().min(1),
-        numberValue: z.number().int().optional(),
+        numberValue: z.number().optional(),
         textValue: z.string().optional(),
       })
     )
